Restructure: Move sanity code to studio folder
