+++ package.json
@@ -9,8 +9,10 @@
     "lint": "next lint",
     "prettier": "prettier '**/*.{js,ts,tsx,css,md,yml}' --write",
     "lint:prettier": "prettier '**/*.{js,css,md,ts,tsx}' --check",
-    "sanity:typegen": "sanity schema extract && sanity typegen generate"
-  },
+    "sanity:generate-schema": "sanity schema extract --path ./studio/generated/sanity-schema.json",
+    "sanity:generate-types": "sanity typegen generate --config-path ./studio/sanity-typegen.json",
+    "sanity:typegen": "yarn sanity:generate-schema && yarn sanity:generate-types"
+    },
   "dependencies": {
     "@portabletext/react": "^3.0.2",
     "@sanity/client": "^7.2.1",
