+++ pages/masterfadderperioden.tsx
@@ -8,10 +8,10 @@ import { ApiEvent } from "@/utils/types";
 import Link from "@/components/Link";
 import { groq } from "next-sanity";
 import { TypedObject } from "sanity";
-import { FPGroups } from "@/schemas/dayDescription";
+import { FPGroups } from "@/studio/schemas/dayDescription";
 import { FACEBOOK_GROUP_FOURTHYEARS, MIDT, MSTCNNS } from "@/utils/constants";
-import { sanityClient } from "@/utils/sanity";
-import getSettings, { BlacklistType, Settings } from "@/utils/settings";
+import { sanityClient } from "@/studio/api/client";
+import getSettings, { BlacklistType, Settings } from "@/studio/api/settings";
 
 export type DayDescription = {
   date: string;
