+++ studio/sanityImageUrl.ts
@@ -1,6 +1,6 @@
 import imageUrlBuilder from '@sanity/image-url'
 import type { SanityImageSource } from '@sanity/image-url/lib/types/types'
-import { sanityClient } from './sanity'
+import { sanityClient } from './api/client'
 
 const builder = imageUrlBuilder(sanityClient)
