+++ studio/schemas/dayDescription.ts
