+++ studio/schemas/index.ts
@@ -2,4 +2,9 @@ import { FPSchema, MFPSchema } from "./dayDescription";
 import SiteSettingsSchema from "./siteSettings";
 import TaskforceSchema from "./taskforce";
 
-export const schemaTypes = [FPSchema, MFPSchema, SiteSettingsSchema, TaskforceSchema];
+export const schemaTypes = [
+  FPSchema,
+  MFPSchema,
+  SiteSettingsSchema,
+  TaskforceSchema,
+];
