+++ studio/schemas/siteSettings.ts
